Extract CardWrapper component
